+++ apps/sapa-dev/pages/main/reload-fragment/Second.jsx
@@ -1,0 +1,3 @@
+export function Second() {
+  return <div class="second">Second 3 4</div>;
+}