+++ apps/sapa-dev/pages/main/reload-fragment/start.jsx
@@ -1,0 +1,5 @@
+import { start } from "@elf-framework/sapa";
+
+import { App } from "./App";
+
+start(<App />);