+++ packages/sapa/src/UIElement.js
@@ -206,7 +206,9 @@ export class UIElement extends EventMachine {
 
       // 함수형 컴포넌트는 instance 인지 체크를 해야할 수도 있다.
       isInstanceOf(...args) {
-        return args.some((TargetClass) => NewFunctionComponent === TargetClass);
+        return args.some((TargetClass) => {
+          return NewFunctionComponent === TargetClass;
+        });
       }
       template() {
         return NewFunctionComponent.call(this, this.props);