+++ packages/base-editor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @elf-framework/base-editor
 
+## 0.0.244
+
+### Patch Changes
+
+- refactoring reconcile
+
+- Updated dependencies []:
+  - @elf-framework/sapa@0.0.244
+  - @elf-framework/ui@0.0.244
+
 ## 0.0.240
 
 ### Patch Changes